+++ src/Inventory.Application/Products/CreateNewProduct/CreateNewProductCommand.cs
@@ -3,4 +3,5 @@
 namespace Inventory.Application.Products.CreateNewProduct;
 
 public record CreateNewProductCommand(string Name) : CommandBase;
-public record CreateNewProductCommandResult(string Name) : CommandBase<string>;
+
+public record CreateNewProductWithResultCommand(string Name) : CommandBase<string>;
