+++ src/Inventory.Infrastructure/AutofacModules/ProcessingModule.cs
@@ -20,12 +20,10 @@ public sealed class ProcessingModule : Module
 
         builder.RegisterGeneric(typeof(ValidationRequestPreProcessor<>)).As(typeof(IRequestPreProcessor<>));
 
-        builder.RegisterGenericDecorator(typeof(UnitOfWorkCommandHandlerDecorator<>), typeof(IRequestHandler<>));
-        builder.RegisterGenericDecorator(typeof(UnitOfWorkCommandHandlerWithResultDecorator<,>), typeof(IRequestHandler<,>));
+        builder.RegisterGenericDecorator(typeof(UnitOfWorkCommandHandlerDecorator<,>), typeof(IRequestHandler<,>));
 
         builder.RegisterGenericDecorator(typeof(DiagnosticQueryHandlerDecorator<,>), typeof(IRequestHandler<,>));
 
-        builder.RegisterGenericDecorator(typeof(LoggingRequestHandlerDecorator<>), typeof(IRequestHandler<>));
-        builder.RegisterGenericDecorator(typeof(LoggingRequestHandlerWithResultDecorator<,>), typeof(IRequestHandler<,>));
+        builder.RegisterGenericDecorator(typeof(LoggingRequestHandlerDecorator<,>), typeof(IRequestHandler<,>));
     }
 }
