+++ src/Inventory.Infrastructure/Decorators/UnitOfWork/UnitOfWorkCommandHandlerDecorator.cs
@@ -4,24 +4,24 @@ using MediatR;
 
 namespace Inventory.Infrastructure.Decorators.UnitOfWork;
 
-internal sealed class UnitOfWorkCommandHandlerDecorator<T> : ICommandHandler<T>
-       where T : ICommand
+internal sealed class UnitOfWorkCommandHandlerDecorator<TCommand, TResult> : ICommandHandler<TCommand, TResult>
+        where TCommand : ICommand<TResult>
 {
-    private readonly IRequestHandler<T> _decorated;
+    private readonly IRequestHandler<TCommand, TResult> _decorated;
     private readonly IUnitOfWork _unitOfWork;
 
-    public UnitOfWorkCommandHandlerDecorator(IRequestHandler<T> decorated, IUnitOfWork unitOfWork)
+    public UnitOfWorkCommandHandlerDecorator(IRequestHandler<TCommand, TResult> decorated, IUnitOfWork unitOfWork)
     {
         _decorated = decorated;
         _unitOfWork = unitOfWork;
     }
 
-    public async Task<Unit> Handle(T command, CancellationToken cancellationToken)
+    public async Task<TResult> Handle(TCommand comamnd, CancellationToken cancellationToken)
     {
-        await _decorated.Handle(command, cancellationToken);
+        var result = await _decorated.Handle(comamnd, cancellationToken);
 
         await _unitOfWork.CommitAsync(cancellationToken);
 
-        return Unit.Value;
+        return result;
     }
 }
